Python: Fix PyArrow Type conversion (#6346)
* Python: Fix PyArrow Type conversion

It turns out that we need to cast the literal to the right type:

```
ArrowNotImplementedError: Function 'greater_equal' has no kernel matching input types (timestamp[us, tz=+00:00], int64)
```

After casting the scalar, it works perfect. One awkward thing to note that
we don't have to do this for the `isin` predicate. There this is done
for us. Please check the Jupyter notebook.

* Explicitly set the type