Python: Publish the docs by hand (#6719)
Currently everything that goes into master will directly be pushed
to Github Pages. This was fine for a while, but now we have PRs that
publish new functionality, but then it isn't available in the release
for example: https://github.com/apache/iceberg/pull/6644

I think we should now publish by hand, and move to multi version docs
once PyIceberg stabilizes a bit.