Python: Implement Cython based decoding for Avro (#8134)
* Python: Improve Avro read performance

Utilize __slots__ for various Avro classes.

Memoize the key/value reader functions in MapReader.

* Python: improve Avro parsing by caching struct types.

Cache the struct types for ManifestEntry and ManifestFile schemas.

Cache the _position_to_field_name lookup in Records if a struct type is passed.

Improve StructReader to use slots and avoid isinstance() lookups on each time
the structure is read, by performing the check in __init__().

Improve StructReader to determine the arguments to create_struct() once in init
and reuse the result each time the structure is read, avoid a try block.

* Python: Improve Avro Parse speed

Cache the hash value of the Struct types so that when new
Record objects are created time is not spent recomputing
the hash value.

Change the implementation of the Record class to cache the
mapping from field name to position.  Also utilize an array
rather than a dictionary for these lookups.

* fix some lints

* fix some type casing

* fix: adjust PR comments

* WIP: changes for cython parsing

* fix: address PR feedback.

* Add additional tests for InMemoryBinaryDecoder

* More work on fast Avro decoder

* Work in progress

* Load entire Avro file into memory rather than streaming it

* fix mypy and lint warnings

* fix some more pylint warnings

* Fix lints and build failures

* fix: move lazydict out to its own file.

* fix: disable seeking on input file

* Cleanup expression

* fix: improve integer test case

* fix: rename decoding functions

* fix: remove reader functions that returned native types

* fix: simplify some ReaderClasses via inheritance

Numerous reader classes are simply reading integers, they can all
inherit from IntegerReader rather than having their own implementations
of read() and skip().  Make this change.

* fix: make sqlalchemy optional

* fix: make lazydict generic and move out of avro, add tests

* fix: some lint fixes

* fix: some cleanups and comments

* fix: commit linter changes

---------

Co-authored-by: Fokko Driesprong <fokko@apache.org>