Arrow: Set field-id with prefix
While working on the write support, I started to understand the
meaning of the Arrow metadata. The `PARQUET:` prefix means that
it is specific to Parquet, and setting `PARQUET:field_id` will
actually not be set as metadata, but as the `field_id`. I think
we should remove the other option without the prefix.

I also did the same for the doc. I checked, in Java we don't check
for the documentation, but I think it is a nice addition. However,
I think we should only check for the key that's the same as the
Iceberg field (`doc`).