Add support for `write.metadata.path` (#1642)
Adding support for writing metadata to a custom path set via
`write.metadata.path` property. Since the Python library consolidates
the table operation classes in both the table and catalog classes, I had
to surface the metadata file location handling to the base `Catalog`
class to avoid circular dependencies. This way we are also able to
centralize the metadata location handling for table metadata and
snapshots.


Relates to #1492

---------

Co-authored-by: Kevin Liu <kevinjqliu@users.noreply.github.com>
Co-authored-by: Fokko Driesprong <fokko@apache.org>
Co-authored-by: smaheshwar-pltr <maheshwarsreesh@gmail.com>