Avoid reusing shared metrics evaluator (#1664)
Hello!

I have noticed the same issue as Issue #1506 where the number of results
retrieved is inconsistent across reads and traced the issue to the reuse
of the same metrics evaluator across threads when reading manifests.
Because the metrics evaluator is stateful, this will result in the wrong
results being retrieved nondeterministically, depending on the execution
order of the threads.

This PR addresses the issue by creating a single metrics evaluator per
thread, which I have tested locally. Please let me know if there are any
tests I can add, and I am happy to receive feedback.

Thank you!

Closes #1506