REST: Delegate parsing to Pydantic
Right now we deserialize the JSON into a dict, which is then passed
into the Pydantic model. It is better to fully delegate this to
pydantic because it is probably faster, and we can detect when
models are created from json or from Python dicts.

Required by https://github.com/apache/iceberg-python/pull/1770