Refactor `Metadata` in `Transaction`
Today, we have a copy of the `TableMetadata` on the `Table` and
the `Transaction`. This PR changes that logic to re-use the one
on the table, and add the changes to the one on the `Transaction`.

This also allows us to stack changes, for example, to first change
a schema, and then write data with the new schema right away.

Also a prerequisite for https://github.com/apache/iceberg-python/pull/1772