Build: Decouple setup and exec in integration tests for better flexibility (#1985)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change

Decouple setup and exec in integration tests for better flexibility and
isolation.

For example, currently, `make test-integration` must be executed in a
physical server.
It is because `docker xxx` cannot be executed within a container (docker
in docker is out of scope of this PR).
Sometimes, it is useful for local debugging, as poetry cannot provide
system level isolation, includes environment variables and shared libs.

This PR allows us to run `make test-integration-exec` within a container
environment, after we have run `make test-integration-setup` on this
server.

# Are these changes tested?
Existing tests.

# Are there any user-facing changes?

No.