fix: add metadata_properties to _construct_parameters when update hive table (#2013)
<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->
Closes: https://github.com/apache/iceberg-python/issues/2010

# Rationale for this change
This change adds metadata_properties to the _construct_parameters
function to ensure metadata properties are included in the parameters.
I'm not entirely confident about the changes, so please let me know if
my understanding is correct—if so, I’ll proceed to add tests.
Thanks you!

# Are these changes tested?
Not yet.

# Are there any user-facing changes?
Not sure.