Fix support for writing to nested field partition (#2204)
Closes #2095 

# Rationale for this change
Currently, we can only partition on top-level valid field types, but
this PR adds support for partitioning on primitive fields in a struct
type using dot notation to determine the partitions against the nested
structure.

# Are these changes tested?
Yes added tests and tested a write against the problem in the above
issue.

```
> aws s3 ls s3://myBucket/demo1/nestedPartition/data/
                           PRE timestamp_hour=2024-01-15-10/
                           PRE timestamp_hour=2024-01-15-11/
                           PRE timestamp_hour=2024-04-15-11/
                           PRE timestamp_hour=2024-05-15-10/
```

# Are there any user-facing changes?

no but now can add data to tables that are partitioned by a source
column that's in a struct