add support to configure anonymous in s3 clients (#2323)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change
Currently, it is not possible to configure `anonymous` in the underlying
`s3fs` client (`False` by default).

Resolves #2126

# Are these changes tested?
Yes, unit tests added to confirm the setting is used when the property
is set.

# Are there any user-facing changes?
Yes, users can now set the property `s3.anonymous` to True to enable. If
not set, the behavior will be the same as it currently is.

<!-- In the case of user-facing changes, please add the changelog label.
-->

---------

Co-authored-by: garrett.weaver <garrett.weaver@cloudkitchens.com>