Hiv: Fix renaming to a table that already exists (#2336)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change

Hive catalog returned a `NoSuchNamespaceError` when it tried to rename
to a table that already exists in the catalog. This change will first
check for the destination table existence before executing the alter
table.

# Are these changes tested?

Yes, added one unit test and one integration test

# Are there any user-facing changes?

Yes? Users using the hive catalog will get a new exception for this edge
case.

<!-- In the case of user-facing changes, please add the changelog label.
-->