Fix: use new snapshot id in deleted manifest entry unless is existing entry (#2266)
<!--
Thanks for opening a pull request!
-->
<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change

Based on iceberg spec, when a manifest entry is marked as deleted, the
snapshot id for when the entry was deleted should be used.
https://iceberg.apache.org/spec/?h=deletes#manifest-entry-fields
```
When a file is replaced or deleted from the dataset, its manifest entry fields store the snapshot ID in which the file was deleted and status 2 (deleted). The file may be deleted from the file system when the snapshot in which it was deleted is garbage collected, assuming that older snapshots have also been garbage collected [1].
```


https://github.com/apache/iceberg/blob/main/core/src/main/java/org/apache/iceberg/ManifestWriter.java#L178-L179

Incorrect snapshot id could lead to data being deleted during garbage
collection when not supposed to.

# Are these changes tested?

Added in `test_deletes.py`

# Are there any user-facing changes?

<!-- In the case of user-facing changes, please add the changelog label.
-->

No

---------

Co-authored-by: Leon Lin <lliangyu@amazon.com>
Co-authored-by: Fokko Driesprong <fokko@apache.org>