Replace pre-commit by prek (#2533)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change

It replaces pre-commit with prek, which is almost 10x faster. prek is a
port of pre-commit to rust.

## Are these changes tested?

Yes. `make lint` is running locally 

## Are there any user-facing changes?

Only for developers, given that it affects `make lint`,

**NOTE**: I had to update the poetry.lock file with `poetry lock`,
resulting in multiple changes. Maybe it's better to run again in your
machine (I don't know if it's run automatically in the CI pipeline).

<!-- In the case of user-facing changes, please add the changelog label.
-->