[Upstream Error] Pin pydantic version <2.12.0 to avoid regression (#2591)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change
Closes #2590 

I seem to be faster than the dependabot : ). 

Pin pydantic version to `<2.12.0` to avoid the recent regression:
https://github.com/pydantic/pydantic/issues/12347.

I made this an upper bound because it is unclear about the fix plan in
pydantic community. Once their fix is out, we could release the upper
bound and exclude affected versions only

## Are these changes tested?
No dependency version update. Tested the new version in
https://github.com/apache/iceberg-python/pull/2589 and confirmed the
regression

## Are there any user-facing changes?
No

cc: @Fokko @kevinjqliu 

<!-- In the case of user-facing changes, please add the changelog label.
-->