Limit split so colons are allowed in the password (#2706)
<!--
Thanks for opening a pull request!
-->

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change

As of now, pyiceberg will fail if the provided catalog secret contains a
colon.
By limiting the split, this gets fixed.

## Are these changes tested? 

Yes, locally with `make test` (the test was modified to include a colon
in the secret) and also with a live system.

## Are there any user-facing changes?

Not exactly, just a bug fix.

<!-- In the case of user-facing changes, please add the changelog label.
-->

Co-authored-by: Leandro Martelli <leandro@mazehq.com>