Enables linter rule UP006 (#2744)
<!--
Thanks for opening a pull request!
-->
Part of #2700 

<!-- In the case this PR will resolve an issue, please replace
${GITHUB_ISSUE_ID} below with the actual Github issue id. -->
<!-- Closes #${GITHUB_ISSUE_ID} -->

# Rationale for this change
This enables linter rule UP006.

Most of these changes revolve around using native-Python types instead
of using typing imports.

This is a big one, sorry! It seems easier just to do this in a single PR
(considering how mechanical this process is) as opposed to breaking it
up.

## Are these changes tested?
`make lint` and `make test` should pass.

## Are there any user-facing changes?
None.

<!-- In the case of user-facing changes, please add the changelog label.
-->