Remove `Generic` from expressions (#2750)
# Rationale for this change

This is a big one, and was on my mind for a long time, but I think we
should clean this up before a 1.x release.

I bumped into this once more when I tried to extend from the
`IcebergBaseModel`, but that did not work well with the `Generic`.

From the beginning we had these Generic in the expressions system, but
it really never worked as we hoped. It came from Java where Generics are
much stronger, but the static typing of Python/mypy doesn't really
follow the types.


## Are these changes tested?

## Are there any user-facing changes?

<!-- In the case of user-facing changes, please add the changelog label.
-->