feat: Make `and` expression JSON serializable (#2784)
Related to: #2518, #2775 

# Rationale for this change

This work was done by @Aniketsy, I just opened this to get the tests
passing, and we can merge for scan planning.

But, this PR allows `And` expressions to be deserialized from JSON
through Pydantic.

This PR aligns the `And` expression with the `Or`/`Not` pattern by
adding `IcebergBaseModel` as an inherited class. This gets teh And
expression into a proven serializable state, preparing it for the full
expression tree [de]serializability work in #2783.

## Are these changes tested?

Yes added a test and ensure that they align with EpressionParser in
Iceberg Java

## Are there any user-facing changes?

No this is just serialization

cc: @kevinjqliu @Fokko

---------

Co-authored-by: Aniket Singh Yadav <singhyadavaniket43@gmail.com>