Remove `field-id` constraint on `add_files` (#2662)
# Rationale for this change
Closes #2131

The PR relaxes the constraint that prevented adding any file with field
IDs, and replaces it with a constraint that prevents adding files which
contain field IDs that are inconsistent with the field IDs of the table.
If the field IDs are compatible, then they can be added safely, if not,
they will be rejected.

## Are these changes tested?
Yes

## Are there any user-facing changes?
Yes