feat: Add RESTscan Planning Endpoints and config (#2842)
related to #2775 

# Rationale for this change
This PR adds some of the rest scanning endpoints needed for the
synchronous scan planning And the configuration property used to
determine scan planning support.

Aligning with the java implementation in
https://github.com/apache/iceberg/pull/13400.

## Usage

```
catalog = RestCatalog(
 "the-best-rest-catalog", uri="http://127.0.0.1:8181", **{"rest.scan-planning.enabled": "true"},
)
```
Ideally we only would want to use rest scan planning if the catalog is
**_capable_**. I remember there was a PR open for this at one point but
can't seem to find it.

## Are these changes tested?

## Are there any user-facing changes?

new property to configure scanning