feat: Add models for rest scan planning (#2861)
related to #2775 and #2792

# Rationale for this change
This PR adds the Pydantic models for the REST catalog server-side scan
planning API focusing on the synchronous use cases first.

There's some redundancy here with things like `RESTDataFile` vs the
existing `DataFile` in the manifest module.
As mentioned in #2792 the manifest logic with encoding/decoding is
dependency on Avro. Rather than trying to solve that unification problem
upfront and blocking ourselves, I went with separate REST pydantic types
for now. The plan is to eventually add conversion methods or a common
interface so these can work together with our existing scan task and
content file types.

The models and tests here align with the Java implementation.

## Are these changes tested?

Yes, added tests and works against POC

## Are there any user-facing changes?

No