feat: Allow servers to express supported endpoints with ConfigResponse (#2848)
closes to #2847 

# Rationale for this change

This PR adds the server endpoint capabilities support, aligning with the
Java
[implementation](https://github.com/apache/iceberg/blob/main/core/src/main/java/org/apache/iceberg/rest/RESTSessionCatalog.java).
While working on the REST scanning support, we need to know if a server
supports specific capabilities before making any calls. So this PR also
adds some extra support for the current implementation of PI iceberg
REST catalog.

The REST catalog will now parse the endpoints field from the config call
to determine server capabilities. When a server doesn't respond, we have
fallback logic that matches the behavior of Java's rest catalog. The
View endpoints are conditionally added to the default with the config
property as well.

## Are these changes tested?

Added unit tests and tested with the iceberg rest fixture.

## Are there any user-facing changes?
Yes added config and alignment with java impl.

cc: @kevinjqliu @Fokko